remove redundant files and moves all commands into main.rb
